Updated initial files.
This update includes the whole unity files of this project.
Currently, it works for three levels of Tower of Hanoi, but there're still bugs/glitches that needs to be fixed in the future.
Large files were not pushed onto Github due to its 100MB commit limit. Therefore, the assets and Library were stored in my personal outlook drive. I believe that the important parts in this project is its scripts and how it works in VR.